api/requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
iniconfig==1.0.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
openapi-codec==1.3.2
packaging==20.4
Pint==0.14
pluggy==0.13.1
py==1.9.0
pylint==2.5.3
pyparsing==2.4.7
pytest==6.0.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
unit-converter==1.1.0
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
